Move classes about self-contained JAR files to org.embulk.cli
