+++ embulk-core/src/main/java/org/embulk/cli/Main.java
@@ -7,7 +7,6 @@ import java.util.concurrent.LinkedBlockingQueue;
 import java.util.jar.Manifest;
 import org.embulk.EmbulkDependencyClassLoader;
 import org.embulk.EmbulkSystemProperties;
-import org.embulk.deps.EmbulkSelfContainedJarFiles;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.slf4j.event.SubstituteLoggingEvent;
@@ -18,7 +17,7 @@ public class Main {
         final Manifest manifest = CliManifest.getManifest();
 
         // They are loaded before SLF4J is initialized along with Logback. They don't use SLF4J for error logging.
-        EmbulkSelfContainedJarFiles.staticInitializer().addFromManifest(manifest).initialize();
+        SelfContainedJarFiles.staticInitializer().addFromManifest(manifest).initialize();
         EmbulkDependencyClassLoader.staticInitializer().useSelfContainedJarFiles().initialize();
 
         // Using SubstituteLogger here because SLF4J and Logback are initialized later (CliLogbackConfigurator.configure).
