+++ embulk-core/src/main/java/org/embulk/cli/Resource.java
@@ -1,4 +1,4 @@
-package org.embulk.deps;
+package org.embulk.cli;
 
 import java.net.MalformedURLException;
 import java.net.URL;
